// generated/api-client/models/facility-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface FacilityResponseDto
 */
export interface FacilityResponseDto {
    /**
     * Facility ID
     * @type {string}
     * @memberof FacilityResponseDto
     */
    'id': string;
    /**
     * Facility name
     * @type {string}
     * @memberof FacilityResponseDto
     */
    'name': string;
    /**
     * Facility description
     * @type {string}
     * @memberof FacilityResponseDto
     */
    'description': string;
    /**
     * Facility location
     * @type {string}
     * @memberof FacilityResponseDto
     */
    'location': string;
    /**
     * Facility capacity
     * @type {number}
     * @memberof FacilityResponseDto
     */
    'capacity': number;
    /**
     * Current occupancy count
     * @type {number}
     * @memberof FacilityResponseDto
     */
    'currentOccupancy': number;
    /**
     * Whether facility is active
     * @type {boolean}
     * @memberof FacilityResponseDto
     */
    'isActive': boolean;
    /**
     * Facility creation date
     * @type {string}
     * @memberof FacilityResponseDto
     */
    'createdAt': string;
    /**
     * Facility last update date
     * @type {string}
     * @memberof FacilityResponseDto
     */
    'updatedAt': string;
}